+++ docs/src/release_notes.rst
@@ -15,6 +15,13 @@ Release Notes
 0.8.x
 =====
 
+0.8.27
+------
++ Added decorator to provide common methods for Playbook Apps.
++ Added :py:meth:`~tcex.tcex_playbook.TcExPlaybook.add_output` and :py:meth:`~tcex.tcex_playbook.TcExPlaybook.write_output` methods to provide an alternative way to write playbook output data.
++ Added access to resolved args.
++ Updated ``tclib`` logic for **lib_latest** symbolic link.
+
 0.8.26
 ------
 + Updated ``tcinit`` to include **migration** as an action to help convert non App Builder compliant Apps.